polish up the dockerfile implementation after experimenting
Add a default configuration which listens to all networks.

This makes the getting started process simpler for users, since the
docker invocation for the image can be as simple as:

docker run -p 4000:4000 \
  --env APOLLO_GRAPH_REF="<your graph>" \
  --env APOLLO_KEY="<your key>" \
  --rm \
  ghcr.io/apollographql/router:v0.1.0-preview.1

This will use the default configuration, use uplink to retrieve graph
details and remove the image when it terminates.