dockerfile example and publish on release to github (#764) (follow-up)
Found bug in dockerfile during release process. The dockerfile copied
the default configuration from the wrong directory. By bad luck, it
found a configuration in the specified location, which meant the build
succeeded, but that config didn't work in docker (wrong listen address
for router).

This fix ensures that the correct config file is included in the docker
image.

We've re-issued the fixed docker files manually.