provide a mechanism for configuring subgraph responses
The mechanism needs to be easily accessible with apollo-router-core so
I've expressed the mechanism as part of the traffic shaping plugin.

Even though this plugin is experimental, it does feel like this is the
right place for this configuration option since it is a form of traffic
shaping.

By default, the reporting of subgraph responses (errors and extensions)
is suppressed. With configuration it may be turned on:

```
plugins:
  experimental.traffic_shaping:
    include_subgraph_responses: true
```

As part of this, I've also migrated the benchmark implementation of a
MockSubgraph service into plugin_utils and fixed (with help from Jeremy
and Geoffroy) the benchmarks.

I'm unsure that the place where I've put the mock subgraph is a good
location in plugin_utils. I noted there are already some mock services
in there and it wasn't obvious how this mock related to the existing
ones. That definitely needs some cleaning up.

Filing this as a draft because I definitely want opinions on:
 - Putting the configuration on the traffic shaping plugin
 - Cleaning up plugin_utils a bit to find a good home for the new
   MockSubgraph service.

fixes: #743