tighten integration with studio usage protobuf definition (#776)
Previously, we checked in a copy of the reports.proto file into the
router repository. This was problematic because the reports.proto file
could change and we might find that the router was out of step with the
source version.

This change removes the checked in copy and updates spaceport build.rs
to try and retrieve the reports.proto file from the canconical location.

If we just wrote the file every time, it would cause re-builds every run
which is slow, so it checks if there is an existing (cached) file
locally and if there is checks the contents to see if there has been a
change and only updates the local (cached) file if there has been a
change.

The fix is a bit fragile, since it makes assumptions about the format of
the proto file and edits the contents appropriately. It would be better,
in future, to be able to avoid the post-processing if possible.

Updated .gitignore to make sure no-one checks in the cached copy by
mistake.

fixes: #423