introduce new experimental plugin to allow subgraph error responses (#797)
* introduce new experimental plugin to allow subgraph error responses

This works in a similar fashion to the headers plugin. You can configure
all to apply to all subgraphs and/or a subgraphs hashmap which overrides
the all specification on a per subgraph basis.

I've also migrated some of the benchmark mock code into
apollo-router-core since it's helpful for testing this kind of thing and
fixed the benchmarks (with help from @Geal and @o0Ignition0o).