store the changelog for the next release in a separate file (#807)
the NEXT_CHANGELOG.md file holds the changes for the next release. That
way, it can be copied directly by the release script, holding only the
most recent changes, and the CHANGELOG.md can contain the full list of
changes (that is currently commented out due to the release script).