split the CI caches for .cargo and target/
On every code change, the target/ folder will change, so we will upload
a new cache. But a part of the cache, like the .cargo folder and the
node_modules folder for tests, rarely changes. By splitting the caches,
we can make sure only target/ is uploaded in most cases