do not try to generate a JSON value on a "flatten" path
the Value::from_path method is used to generate a json value by
inserting an object at a specific path, so the value has a certain
shape.
If we have { "c": 1 } and want to generate { "a": { "b": { "c": 1 } } },
we will call from_path with the path "/a/b". If we have a path in an
array element, it will contain the index, as follows: "/a/1/b".

In the case of federated queries, for each of the representations, we
generate a path at which the response will be inserted. But in the case
we could not select any representation, we avoid the subgraph query
(since it would return nothing) and return an empty value. This empty
value cannot be null, it has to match the expected shape (otherwise
we'll run into response formatting issues), so we use Value::from_path
with the current path and the null value.

The bug here is that if the current path contains a @ (flatten), which
means that we try to select from all elements of an array, we were
trying to generate the value shape even after the flatten element, while
we have way of knowing if the array there even has elements. In some
cases the array was even replaced by an object afterwards.
So here the safest solution is to avoid generating anything once we
reach the flatten element, and return the partially generated shape