Switch from warp to axum  (#751)
This PR enables by default the use of axum instead of warp. I added 2 features in Cargo.toml to switch from axum to warp if needed.

I spotted different behaviors between warp and axum. Indeed, Axum is stricter than warp, if you don't specify the Content-Type: application/json in the request headers, axum will throw an error with status code 415. Another difference is when we're using unix socket, before if it was a GET or a POST warp accepted the query directly in the body of the request. Now with axum if it's a GET it needs to be in query parameters and not in the body of the request..

I also introduced a small fix when introspection is disabled, no need to go further and ask something to the query planner, we are now making an early return with an error

Signed-off-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>

Co-authored-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>