update buildstructor to 0.1.5 and start converting all tests
Now that the buildstructor is working, use it for all Request/Restore
structures and start updating all the tests to make use of it.

This is time consuming and difficult/boring work. Many of the existing
tests don't fully create the required structures, so I'm having to add a
lot of extra details to get this to work.

Might be easier if I start providing defaults for various common
components.

About 1/2 the tests are converted at this point with about 950 lines of
output errors remaining.