Fix telemetry shutdown in the case of a failed config update. (#849)
* Fix telemetry shutdown in the case of a failed config update.
The root cause is that tracing providers MUST be flushed before drop. otherwise they hang. This is really nasty, but the fix needs to happen in Otel.

Flushing will happen automagically if a provider is set as the global for otel.

In the case of a plugin after the otel plugin failing to start up correctly, or even one of the subsequent methods inside the otel plugin fails then the provider is not set as the global and we must flush it.

Flushing must occur in a separate thread.

Flushing is now moved to the dropping of a telemetry plugin.

* add a regression test to the telemetry hang

* Update apollo-router/src/testdata/invalid_supergraph.graphql

Co-authored-by: Jeremy Lempereur <jeremy.lempereur@gmail.com>

Co-authored-by: bryn <bryn@apollographql.com>
Co-authored-by: o0Ignition0o <jeremy.lempereur@gmail.com>