Fix the bug which impacted apq tests and sundry cleanup
I couldn't make the buildstructor on request change stick. Too
complicated just now since we use many features on the TypedBuilder.

I used the get rid of setter(strip-option) param. I also made data a
required field on the RouterResponse builder because it's required in
the Response structure and it's safest to match.

Re-enable the APQ tests and replace use of constructors in various
places with builders (cosmetic, but nice to be neat).