set rust toolchaing back to 1.59.0
For some reason, since I merged with main this morning, having a rust
toolchain for 1.60.0 causes the link step of build to run out of memory
on the circle-ci linux executor. That's a really odd one, because I
wouldn't say the merge contained any odd things which would trigger this
problem.

For now, set it back to 1.59.0, since manual ssh testing on circleci
shows that works.