fix: extract operation name correctly from a get query.
In a get query, the operation name was extracted as if it had been a JSON payload.
We misstook this bug for an instrument related issue (we thought instrumentation had a bug when it wouldnt display the operation name).

this PR adds http get and http post tests to make sure query and operation name will always be extracted correctly from the http server.