go paste crazy with macros from hell
Most of the stuff we do in the rhai crate is just slight variations on
the same them for different types. I probably could use generics/traits
to get to a "proper" answer, but my head is hurting and every now and
then I miss Lisp and the CPP. So, I've used "paste" to rapidly prototype
the ideas around rhai plugin participation.

I've got a bunch of macro_rules that generate code based around a basic
ident (e.g. router, query_planner) and then fill in a skeleton.