do not remove __typename from input values
to avoid copies, response formatting removes element from the input
value to add them to the output value. In the case of __typename though,
if it is explicitely requested by the client, it should not be removed
because it might be used in a fragment later.

Example:

query  {
  get {
    __typename
    ... on Product {
      symbol
    }
  }
}

if __typename is removed from input values, it won't be usable in the
fragment spread to check that the type is Product