use the multi_thread executor to run the telemetry hang test
The test wasn't using the multi_thread executor, so eventually it would
trip over the bug in opentelemetry which we've hit in the past where the
runtime would hang for not fully understood reasons.

see: https://github.com/open-telemetry/opentelemetry-rust/issues/536

Using the multi_thread executor is the recommended work-around.

I've also remove the redundant thread spawn from the telemetry Drop
implementation, since I don't think it's required as well as the
spawn_blocking() call.

I've run this test in a tight loop for a couple of hours (on my M1
laptop) with no issues.