add a test which panics in a mock subgraph and handles the response (#918)
Fix the tests affected by the use of `buffer()` functionality in tower (which leads to lost panics). Also add a new test to the query planner module which illustrates the issue and has a documentation comment describing what is happening.

This doesn't resolve the more deep seated problem of "how do we know when a test is affected by this problem". That's a longer term issue to wrestle with.