proper behaviour for the response data field (#926)
* proper behaviour for the response data field

According to the GraphQL specification (
https://spec.graphql.org/October2021/#sec-Response ), the `data` field
in the response can have varying behaviours:

> If an error was raised before execution begins, the data entry should
not be present in the result.

> If an error was raised during the execution that prevented a valid
response, the data entry in the response should be null.

Following this, we make the data field optional, and set it to
Some(Value::Null) on errors in query execution.

Looking at this also means we should handle that format from subgraphs

* changelog

* fix examples compilation

* remove the test for invalid operation names

invalid operation names are rejected much earlier in the query planner

* deserialize responses correctly

we were converting nulls to None, now they're deserialized correctly azs
Some(Value::Null)