significantly improve the error handling story
We have no way to report errors during service insertion into the
pipeline, so simply log them as errors.

Always use checkpoint service in preference to map_request, because we
can abort the execution pipeline if a rhai map_request fails. We still
call the function map_request in rhai, but under the covers it is using
checkpoint.

Generate and return error responses for failed map_request and
map_response. Ideally, we would use the existing responses for
map_response failures and just append the error to the existing errors.
However, that is complicated due to the variations in the types and the
fact that QueryPlannerResponses don't even have a way to propagate an
error (as far as I can see).

Summary: Things are better, but some way from perfect.