accept endpoint URL without scheme for telemetry
endpoint configuration for Datadog and OTLP take a URL as argument, but
it is easy to make the mistake of setting a value of the format
"host:port". In that case, the URL parser will happily understand that
as "host" scheme and "port" path.
THis PR instead transforms it to "http://host:port"