ensure global scope is preserved for callback functions
The callback code used a function which didn't pass engine scope, so
global scope was not available.

I've changed the invocation of callback functions so that global scope
is shared across invocations and accessible from callbacks.

As part of this I introduced a global constant "apollo_start" which can
be used as a start of time timestamp and added an example which shows
how to use it.