Better diagnostics when integration tests don't have access to gateway (#994)
Quite often, the integration tests will fail because they don't have
access to a gateway instance on which to base their "expected" results.

This modification tries to make this clearer. It's not really a solution
for failing tests, but should at least help us identify the reason
sooner in future.