Reporting followups (#1020)
Spaceport now just deals with reports and forwards them on. There is no aggregation taking place in spaceport.
The telemetry plugin now composes the complete report for studio.
Metrics are aggregated on Key which is now an enum of excluded or regular. Excluded has no information.
Report header is introduced so that we can eventually populate it with all the right info.


Co-authored-by: bryn <bryn@apollographql.com>