use vanity URL for download instructions
This allows us to own the host that is used for installing the router and
mitigate outages if GitHub goes down.  For now, the URL just redirects to
the previous URL, but in the future we could have it try other locations in
the event of an outage.

It also looks much nicer!

Closes #962