fix @inaccessible usage on object types
if @inaccessible is set on an object type, a client should not be able
to receive a value of that type. Instead, it should be replaced with a
null, without an error.

An inaccessible type will not appear in the API schema that we use for
response formatting, so when we get an object with a __typename field,
we check if that typ ename is known in the schema's object types