Fix error count cache counts.
This starts to make the internal rust datastructures deviate from the protobuf. The reason we need to do this is that some of the protobuf fields are derived from the histogram count.

Where DurationHistogram is used I have tried to eliminate the corresponding count fields.
However, it's sometimes not clear from the protobuf which these are.

Finally, DurationHistogram now takes an option of duration. None is ignored.