Fix the deployment template to support not specifying prometheus
The best practice fix appears to be to ensure that elements are present
in the configuration or to check each element in the extended dot
notation reference individually.

Usually, I would go with ensuring we have false elements defaulted in
the configuration, but since this is a complex configuration file I'm
going with keeping the configuration file default clean and having
separate checks.