fix: Remove multi thread flavor for tests that use test span.
This is a workaround until the root cause is adressed.

TLDR: Test spans filters for relevant root nodes to generate the span sequences.
This doesnt play well with multi threading since each thread has a separate root span.
I ll open a deeper issue in test-span, with a couple of steps to fix this.