use subgraphs deployed on Fly in CI (#1090)
* update the deployed supergraph schema

the subgraphs deployed on fly.io for demo purpose were updated, and the
supergraph schema changed

* use the subgraphs and gateway deployed on starstuff in tests

the CI includes complicated steps to start node subgraphs for integration tests. We can instead use already deployed subgraphs.

* fix: Remove multi thread flavor for tests that use test span.

This is a workaround until the root cause is adressed.

TLDR: Test spans filters for relevant root nodes to generate the span sequences.
This doesnt play well with multi threading since each thread has a separate root span.
I ll open a deeper issue in test-span, with a couple of steps to fix this.

* use a local subgraph in jaeger tests

for the jaeger trace propagation test, we cannot use the fly subgraphs
because they cannot reach the local jaeger server, and deploying a
public jaeger would be harder to secure and setup in CI.
Instead, just for this test, we create a local server emulating the
subgraph, that supports jaeger trace propagation. This has the added
benefit that if the header is not sent, we can immediately fail the test

* no need to compile and run xtask for tests now

the tests can be run directly

Co-authored-by: o0Ignition0o <jeremy.lempereur@gmail.com>