CORS mis-configuration shouldn't cause a router panic (#1197)
* CORS mis-configuration shouldn't cause a router panic

Tower validates supplied router configuration and `asserts` that it is
valid to use. If it isn't, this results in a runtime panic which is not
a nice user experience.

This set of changes checks the CORS parameters before passing them to
tower and results in a manageable error if mis-configuration is
detected. It would be better if this was fixed in tower, but I imagine
there is a good reason to have the asserts. I'll file a ticket there
anyway, but in the meantime we can just have this function in our code.