fix: Input value validation.
fixes #1203

The graphql specification provides two sets of coercion / validation rules, depending on whether we're dealing with inputs or outputs.
The spec we were following for query validation used the output coercion rules; which don't match the spec.

This Pull Request renames validate_value to validate_input_value, and changes the tests and the validation mechanism to match the specification on scalars.

I have checked:
- String
- Int
- Id
- Float
- Boolean

Which leaves us with Lists, NonNull and Named.