Convert examples and scaffold to use the new checkpoint_async()
This was a lot harder than I hoped it would be. It took longer to
convert the example than the implementation and the tests.

Also: I spotted a gotcha with the scaffold stuff. Since we don't compile
the scaffold code, there's a very significant chance that we will make
API changes which aren't reflected in our scaffold templates. That's
something we are going to need to watch out for.