Pin clap dependency in Cargo.toml
Workaround against #1231

A minor release of Clap occured yesterday; which introduced a breaking change.

This might lead cargo scaffold users to hit a panic a runtime when the router tries to parse env variables and arguments.

This patch Pins the clap dependency to the version that was available before the release, until the root cause is found and fixed.