create the ExecutionResponse after the primary response was generated
The `@defer` preliminary work has a surprising side effect: when using
methods like `RouterResponse::map_response`, they are
executed before the subgraph responses are received, because they work
on the stream of responses.
This commit goes back to the previous behaviour by awaiting the primary
response before creating the ExecutionResponse.