use the async CacheStorage in APQ
here we can see that we veer off the happy path, because having an async
cache for APQ means that we must first query that cache asynchronously
then depending on the result call the underlying service. For this we
need to make the underlying service clonable so it can be moved into the
future. We make it clonable using tower's Buffer with the same default
value as is used everywhere else