WIP: generate new Service instances on every request
The router relies heavily on tower::Buffer to make services clonable, so
they can be reused for multiples requests, but this is far from the
preferred model of tower services where a new Service instance is
generated for every new connection or channel.

This commit experiments with replacing the RouterService Buffer with a
RouterServiceFactory that generates a new RouterService dynamically.
For now it works without plugins, but they will be added back later