Remove non-GraphQL body variants from RouterResponse
Fixes https://github.com/apollographql/router/issues/1140

The body type inside `RouterResponse` is now `graphql::Response` instead of `ResponseBody`.

`ResponseBody` is an enum for either `graphql::Response`, a JSON value, or a String.
However except for one unit test only the GraphQL variant was ever used.

The diff is big but (again except for that test) the only meaningful change is:

```diff
--- a/apollo-router/src/services/mod.rs
+++ b/apollo-router/src/services/mod.rs
 /// [`Context`] and [`http_ext::Response<ResponseBody>`] for the response.
 ///
 /// This consists of the response body and the context.
 #[derive(Clone, Debug)]
-pub struct RouterResponse<T: Stream<Item = ResponseBody>> {
+pub struct RouterResponse<T: Stream<Item = Response>> {
     pub response: http_ext::Response<T>,
     pub context: Context,
 }
```

The test in question is `it_checks_the_shape_of_router_request`
where the response body was changed from a single JSON string
to a GraphQL response with that same string in `data`.

`BodyResponse` is still used for custom endpoints in plugins
(which is where non-GraphQL responses are used),
but that part of the API does not use `RouterResponse` so it is not affected by this PR.

Everything else was made with search-and-replace or following compiler errors,
and so should be fairly mechanical.