rustfmt: per-item import granularity and std/external/crate grouping
Since this configuration is unstable, it is passed through the command line
to rustfmt Nightly. Having it in `rustfmt.toml` would cause noisy warnings
for folks running rustfmt with a stable toolchain.

The default behavior (used when running rustfmt outside of xtask)
is imports_granularity=Preserve and group_imports=Preserve,
which does not revert any changes other values might do.