attempt at a cache dedup layer
this works similarly to the current cache query coalescing code, but
invert the responsibility: instead of a cache that performs a
get_or_insert with a predefined resolver, the cache returns an Entry
(a bit like HashMap's Entry type) that can be awaited on. That way it
can adapt to use cases like subgraph entity caching.

If there was an element in cache, it will contain the value.
If not, it creates an element in a wait map for the key containing a
broadcast receiver, and returns the Entry to the client. The client
can then perform any required task, and then set the value in the
entry, which will set the value in the cache and notify the receivers