improve plugin initialization
Prompted by Bryn's previous comment about ordering, I looked further
into the plugin registration process and discovered bugs (which caused
plugins not to be ordered) and, in discussion with Bryn, identified
opportunities to improve plugin registration and initialization.

Thus:

 - mandatory plugins are now handled during router factory creation
 - order of plugins is now maintained (and logged)
 - enforced ordering of mandatory plugins

Previously we tried to maintain plugin sequence up to a certain point
and then we inserted them into a HashMap, which meant plugin sequencing
was unpredictable.

Sample log line:

2022-06-29T07:34:53.114688Z  INFO apollo_router::router_factory: list of plugins plugin_details=[("apollo.telemetry", "apollo_router::plugins::telemetry::Telemetry"), ("experimental.include_subgraph_errors", "apollo_router::plugins::include_subgraph_errors::IncludeSubgraphErrors"), ("apollo.csrf", "apollo_router::plugins::csrf::Csrf"), ("example.context_data", "context_data::context_data::ContextData")]

Note that plugins are listed in processing sequence.