Improve configuration validation and environment expansion
Environment expansion now covers the entire configuration file, and supports non string types.
This means that it is not possible to use environment variable in the server section of the yaml, and also in numeric and boolean fields.

Environment variables will always be shown in their original form in error messages preventing leakage of secrets.

These changes allow more of the configuration file to be validated via json schema, as previously we just skipped errors where fields contained env variables.