renovate rustc globally, rather than individually
I observed https://github.com/apollographql/router/pull/1333 was a PR that
Renovate opened to update to a more recent version of `rustc`.  While
updating our Rust version in general makes sense, it doesn't make any sense
to do in isolation.  Rather, it need to be set in various places in our
repository that rely on that being set in lock-step.

Those different places we use the version are not done in way that can
easily be DRY, but Renovate can help us open PRs that touch the right
places all at once.

This uses a special syntax and ... a regex ... to do that.

Renovate is actually built on Regex for the most part, so mostly just
perpetuating it here, I swear.

I chose to use Docker as the version source since I figure once it gets
published to Docker, the version is also available through Rust directly,
and since once of the versions we depend on is a Docker Hub version itself.