chore: update rhai to latest release (1.8.0)
We had been using a pinned git version to pull in a fix which geoffroy
had contributed.

I ran the full test suite and all the examples and all looks good.

I also updated one of the examples based on improvements I identified in
writing rhai script for another project.