experiment with deferred execution
we need to pass the response sender to every recursive call.
Each deferred call must move the needed parts of the query plan into a
spawned task that can send the next response. We create channels thatare
used to aggregate data from fetch nodes and send it to the deferred task