renovate: correct rustc regex to not use line-boundaries
Follows up on https://github.com/apollographql/router/pull/1336

The regex I used in the last PR turned out to be incorrect because the
(m)ultiline flag is not passed to the Renovate Regex-invocation.  Instead,
the match is processed against a "joined" string of the file contents.

I've also added a couple more files which were meant to be grouped into this
originally, now that I'm more confident in the behavior after reviewing the
[Renovate dashboard] and specified a branch name explicitly, to avoid the
`docker-` prefixing which was somewhat counter-intuiative.

[Renovate dashboard]: https://app.renovatebot.com/