reduce poll_ready calls in query deduplication
In its initial version, query deduplication was a layer directly above
the subgraph service, which has no requirements on `Service::poll_ready` calls.
Since its integration in the traffic shaping plugin, deduplication is
now a layer above a buffer above possibly other plugins then the
subgraph service.
The Buffer service reserves slots in `poll_ready()` that are consumed when
`call()` is used, and if `call()` is not used, those slots can take a
long time to drop and be reusable.
Since the query deduplication will prevent a lot of calls to the inner
service, it should not forward its `poll_ready` calls directly to the
underlying buffer, because that would hold a lot of slots.
Instead, this commit makes the query deduplication service return ok to
every `poll_ready` call, then uses `oneshot` or `ready_oneshot` on the
underlying service when the actual call is performed