build different response if defer is used
When we are sure there will be no deferred responses (ie no `@defer`
usage in the query), we can create a response that is sent directly,
without multipart.
When there are deferred nodes in the query plan, we cannot know in
advance how many deferred responses can appear, because some of them
might be canceled if required data from previously executed fetch nodes
is absent. So every deferred response has the `hasNext` field set to
true, and we end the stream with an empty response with `hasNext` set to
false