expose the custom endpoints from RouterServiceFactory
the StateMachine was holding a copy of the plugin list to extract the
custom endpoints registered by some of the plugins, like the prometheus
endpoint.

With the Tower refactoring, that plugin list was empty so the endpoints
were not registered.

This commit makes sure that the plugins are only held byt the
RouterServiceFactory instance, and that it implements a method to return
the list of endpoints, that can be used by the StateMachine instance