rewrite the caching API (#1281)
This introduces a new asynchronous caching API that opens the way to multi level caching (in memory and
database). The API revolves around an `Entry` structure that allows query deduplication and lets the
client decide how to generate the value to cache, instead of a complicated delegate system inside the
cache.

* make the cache module with an async cache storage module

this is a simple wrapper around moka for now, but that uses an async
interface for get and insert, because requesting data from redis later
will be async

* use the async CacheStorage in APQ

here we can see that we veer off the happy path, because having an async
cache for APQ means that we must first query that cache asynchronously
then depending on the result call the underlying service. For this we
need to make the underlying service clonable so it can be moved into the
future. We make it clonable using tower's Buffer with the same default
value as is used everywhere else

* add at a cache dedup layer

this works similarly to the current cache query coalescing code, but
invert the responsibility: instead of a cache that performs a
get_or_insert with a predefined resolver, the cache returns an Entry
(a bit like HashMap's Entry type) that can be awaited on. That way it
can adapt to use cases like subgraph entity caching.

If there was an element in cache, it will contain the value.
If not, it creates an element in a wait map for the key containing a
broadcast receiver, and returns the Entry to the client. The client
can then perform any required task, and then set the value in the
entry, which will set the value in the cache and notify the receivers

* use LruCache instead of moka

this allows a relaxation of trait bounds: the key and value types should
not need to be Sync

* start using in CachingQueryPlanner

We make the cache instantiation async in case it needs to open
connections

* simplify query planner key usage

We introduce a `QueryKey` type instead of the tuple that was passed around

* use AsyncCheckpointService in APQ