Defer: query plan changes (#1182)
* add DeferNode

this is a new kind of query plan node used to support usage of the
@defer directive

* experiment with deferred execution

we need to pass the response sender to every recursive call.
Each deferred call must move the needed parts of the query plan into a
spawned task that can send the next response. We create channels that are
used to aggregate data from fetch nodes and send it to the deferred task

* example deferred usage

* use a router-bridge with a defer aware query planner

* multipart response body encoding

* add hasNext and subselection fields to the graphql response

the subselection field will not be serialized, it is there mainly for
formatting the deferred responses. hasNext is used to indicate if there
will be other responses after the current one

* build different response if defer is used

When we are sure there will be no deferred responses (ie no `@defer`
usage in the query), we can create a response that is sent directly,
without multipart.
When there are deferred nodes in the query plan, we cannot know in
advance how many deferred responses can appear, because some of them
might be canceled if required data from previously executed fetch nodes
is absent. So every deferred response has the `hasNext` field set to
true, and we end the stream with an empty response with `hasNext` set to
false

* add a configuration option for defer

marked experimental and disabled by default for now

* Add support of sub selection formatting for defer (#1398)

add support of subselection in format_response

Signed-off-by: Benjamin Coenen <5719034+bnjjj@users.noreply.github.com>
Co-authored-by: Geoffroy Couprie <geoffroy@apollographql.com>

* format the response depending on the path too

the same subselection could be used in different deferred nodes at
different paths, so response formatting must take it into account.

Since the deferred responses are generated as an object starting at the
root, we use the path to look up selections for formatting but do not
transmit it to the client

* support deferred response labels

* implement ConditionNode

* defining the directive in the supergraph is not necessary anymore

* wait until the primary node is done before sending the deferred response

The deferred node marks dependencies on specific fetch nodes so it can
start right when the necessary is available and execute in parallel of
the primary query. But it must wait for the primary query tto finish
before sending the response, first because it might select data from the
primary query, and second because it will make a better experience to
guarantee the order of primary response then deferred response

* wait for the primary response before starting on the deferred node

this is redundant with the code that only waits for fetches in the
"depends" list, but there might be a way to go back to only those
fetches, unless <some condition> that I have not found yet (maybe the
absence of a subselection in the deferred node).

In the meantime, this solves issues with queries like this:

```
query Query {
  allProducts {
    ... @defer {
      variation {
        id
      }

      ... @defer {
        sku
      }
    }

    identifiant: id
  }
}
```

where all of the data is obtained in one initial subgraph fetch node,
then all the other query plan nodes cherry pick data from it

* wait for the primary response only if necessary

if a node has no depends, it should wait for the primary value, because
it can use it in a subselection, but also can pass it to the inner node

Co-authored-by: Coenen Benjamin <benjamin.coenen@hotmail.com>