Modify plugin API to provide a struct rather than config
This change modifies the plugin API so that the new fn now passes a
structure, PluginInitialise, which contains:
 - Config (generic over plugin config type)
 - Schema (String)

In the future we can add to this structure fairly easily to extend it.
For example, we may add an Apollo Compiler context or replace the
current Schema string with a Compiler context.

Most of the change is updating boiler plate in tests/examples...