remove telemetry configuration hot reloading
Configuration hot reloading is not very useful for telemetry, and is the
source of regular bugs that are hard to fix.

This removes the support for configuration reloading entirely. Now, the
router will reject a configuration reload with an error log if the
telemetry configuration changed.

A temporary tracing subscriber is created when the router is started, and
is used until a subscriber can be created from the configuration and set
as global subscriber. That global subscriber will then be used for all
the requests and will not change for the entire life of the process.