add a way to override the subscriber
When using the router as a library (example: in integration tests), we
might want to use another tracing subscriber than the one created by
default in the telemetry plugin.

This adds a method to the telemetry plugin to directly provide a
subscriber. It will then be modified to use layers for the telemetry
exporters and spaceport, so it won't be possible to deactivate studio
reportingt that way