remove telemetry configuration hot reloading (#1463)
* remove telemetry configuration hot reloading

Configuration hot reloading is not very useful for telemetry, and is the
source of regular bugs that are hard to fix.

This removes the support for configuration reloading entirely. Now, the
router will reject a configuration reload with an error log if the
telemetry configuration changed.

A temporary tracing subscriber is created when the router is started, and
is used until a subscriber can be created from the configuration and set
as global subscriber. That global subscriber will then be used for all
the requests and will not change for the entire life of the process.

* add a way to override the subscriber

When using the router as a library (example: in integration tests), we
might want to use another tracing subscriber than the one created by
default in the telemetry plugin.

This adds a method to the telemetry plugin to directly provide a
subscriber. It will then be modified to use layers for the telemetry
exporters and spaceport, so it won't be possible to deactivate studio
reporting that way