chore(renovate) group `tracing` packages by themselves
Follows up https://github.com/apollographql/router/pull/1502

Putting the `tracing` packages into their own group to allow them to be
updated indepedently from the other packages which are generally more able
to be updated.

As it stands today, we are constrained in our ability to update the
`tracing` packages.

This should make it possible for them to be updated on-demand via the
Renovate Dependency Dashboard, if we choose to do so, and otherwise de-group
them from the other cargo updates.