Disable ftv1 traces reporting on defered queries.
This commit allows to disable ftv1 traces reporting.

It does so by:
Allowing the telemetry plugin to add a do not sample (bool) entry to the context.
Allowing the telemetry plugin to add a do not sample reason to the execution span.

Having those enabled will:
  - Not propagate the ftv1 header to subgraphs.
  - Not create an ftv1 trace
  - log the reason why the trace is not being created in `debug`

The only reason at the moment is "query is defered" but we could add other eventualy.